pyproject.toml
[project]
name = "sportradar-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pymysql>=1.1.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
